ICU-1211 Porting fixes for the KAI compiler.
X-SVN-Rev: 5781

+++ icu4c/source/test/intltest/intltest.cpp
@@ -170,7 +170,7 @@ UnicodeString operator+(const UnicodeString& left, float num)
 // Append a hex string to the target
 UnicodeString&
 IntlTest::appendHex(uint32_t number,
-            int8_t digits,
+            int32_t digits,
             UnicodeString& target)
 {
     static const UChar digitString[] = {
